move fractal_server/user_settings.py into fractal_server/app
